add minor code fixes
change folder enums name to globalEnums, outsource currenciesPairs from
currencies file into its own one
