--- src/utilities/enums/currencies.ts
@@ -1,14 +1,0 @@
-export enum Currencies {
-  USD = 'USD',
-  PLN = 'PLN',
-  GBP = 'GBP',
-  CHF = 'CHF',
-  EUR = 'EUR',
-}
-
-export enum CurrenciesParis {
-  USD_PLN = 'USD_PLN',
-  EUR_PLN = 'EUR_PLN',
-  CHF_PLN = 'CHF_PLN',
-  GBP_PLN = 'GBP_PLN',
-}
