+++ src/utilities/globalEnums/currencies.ts
@@ -1,0 +1,7 @@
+export enum Currencies {
+  USD = 'USD',
+  PLN = 'PLN',
+  GBP = 'GBP',
+  CHF = 'CHF',
+  EUR = 'EUR',
+}
