+++ src/utilities/globalEnums/currenciesPairs.ts
@@ -1,0 +1,6 @@
+export enum CurrenciesPairs {
+  USD_PLN = 'USD_PLN',
+  EUR_PLN = 'EUR_PLN',
+  CHF_PLN = 'CHF_PLN',
+  GBP_PLN = 'GBP_PLN',
+}
